Implement calculation of distance and pace to the right formats | Stop location updates on break

+++ app/src/main/java/gr/dipae/thesisfitnessapp/data/sport/mapper/SportsMapper.kt
@@ -29,7 +29,7 @@ class SportsMapper @Inject constructor() : Mapper {
         return listOf(
             SportParameterRequest(
                 name = "distance",
-                value = distance,
+                value = distance.toLong(),
                 achieved = goalParameter?.type is SportParameterType.Distance && distance <= goalParameter.value
             ),
             SportParameterRequest(
